packages/vision/pyproject.toml
[project]
name = "forensic-vision"
version = "0.1.0"
description = "AI extraction pipeline for FMO"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "pydantic>=2.0",
]
